refactored some code and cleaned it up a bit.
