+++ client/src/components/Auth/TwoFactorScreen.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { useSearchParams, useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { useLocalize } from '~/hooks';
 import { useVerifyTwoFactorTempMutation } from 'librechat-data-provider/react-query';
@@ -9,22 +10,23 @@ type TwoFactorFormInputs = {
   backupCode?: string;
 };
 
-type TwoFactorScreenProps = {
-  tempToken: string;
-  onCancel: () => void;
-};
+const TwoFactorScreen: React.FC = () => {
+  // Get the tempToken from query parameters.
+  const [searchParams] = useSearchParams();
+  const navigate = useNavigate();
+  const tempToken = searchParams.get('tempToken') || '';
 
-const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }) => {
+  // Initialize form, localization, toast, and backup toggle state.
   const { register, handleSubmit, formState: { errors } } = useForm<TwoFactorFormInputs>();
   const localize = useLocalize();
   const { showToast } = useToastContext();
-  // This state controls whether to show the backup code input.
   const [useBackup, setUseBackup] = useState<boolean>(false);
 
+  // Setup the mutation for verifying 2FA.
   const { mutate: verifyTempMutate, isLoading } = useVerifyTwoFactorTempMutation();
 
+  // Handle form submission.
   const onSubmit = (data: TwoFactorFormInputs) => {
-    // Build the payload based on whether the backup code is used.
     const payload: any = { tempToken };
     if (useBackup && data.backupCode) {
       payload.backupCode = data.backupCode;
@@ -34,7 +36,7 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
     verifyTempMutate(payload, {
       onSuccess: (result) => {
         if (result.token) {
-          // On success, redirect or update your auth context as needed.
+          // On successful verification, redirect to home.
           window.location.href = '/';
         }
       },
@@ -45,11 +47,16 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
     });
   };
 
+  // Cancel handler navigates back to the login page.
+  const handleCancel = () => {
+    navigate('/login', { replace: true });
+  };
+
   return (
     <div className="p-4">
       <h2 className="mb-4 text-xl font-semibold">Enter 2FA Code</h2>
       <form onSubmit={handleSubmit(onSubmit)}>
-        {/* Show token input only if not using backup */}
+        {/* Input for the 2FA code if not using backup */}
         {!useBackup && (
           <div className="mb-4">
             <label className="block text-sm font-medium">2FA Code</label>
@@ -59,10 +66,12 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
               placeholder="Enter your 2FA code"
               className="mt-1 block w-full rounded-md border border-gray-300 px-3 py-2"
             />
-            {errors.token && <span className="text-red-500 text-sm">{errors.token.message}</span>}
+            {errors.token && (
+              <span className="text-red-500 text-sm">{errors.token.message}</span>
+            )}
           </div>
         )}
-        {/* Show backup code input only when using backup */}
+        {/* Input for the backup code if using backup */}
         {useBackup && (
           <div className="mb-4">
             <label className="block text-sm font-medium">Backup Code</label>
@@ -72,9 +81,12 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
               placeholder="Enter your backup code"
               className="mt-1 block w-full rounded-md border border-gray-300 px-3 py-2"
             />
-            {errors.backupCode && <span className="text-red-500 text-sm">{errors.backupCode.message}</span>}
+            {errors.backupCode && (
+              <span className="text-red-500 text-sm">{errors.backupCode.message}</span>
+            )}
           </div>
         )}
+        {/* Toggle button between 2FA and backup code */}
         <div className="mb-4">
           {!useBackup ? (
             <button
@@ -94,6 +106,7 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
             </button>
           )}
         </div>
+        {/* Submit and Cancel buttons */}
         <div className="flex items-center justify-between">
           <button
             type="submit"
@@ -104,7 +117,7 @@ const TwoFactorScreen: React.FC<TwoFactorScreenProps> = ({ tempToken, onCancel }
           </button>
           <button
             type="button"
-            onClick={onCancel}
+            onClick={handleCancel}
             className="rounded bg-gray-300 px-4 py-2 text-gray-800 hover:bg-gray-400"
           >
               Cancel
