+++ client/src/components/Auth/index.ts
@@ -4,4 +4,4 @@ export { default as ResetPassword } from './ResetPassword';
 export { default as VerifyEmail } from './VerifyEmail';
 export { default as ApiErrorWatcher } from './ApiErrorWatcher';
 export { default as RequestPasswordReset } from './RequestPasswordReset';
-export { default as TwoFactorScreenWrapper } from './TwoFactorScreenWrapper';
+export { default as TwoFactorScreen } from './TwoFactorScreen';
